feat: 네비게이션을 프로필 페이지로 이동 - 메인페이지에서 하단 네비게이션 제거 - 프로필 페이지에 하단 고정 네비게이션 추가 - 프로필, 친구, 리뷰 링크 및 로그아웃 버튼 포함 - 프로필 페이지에서 현재 탭 하이라이트 표시
